+++ request_api_helper/lib/request.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart';
 
+/// Of `RequestApiHelperData`,`RequestApiHelperDownloadData`
+abstract class RequestApiHelperConfig {}
+
 class FileData {
   List<String> path;
   List<String> requestName;
@@ -8,7 +11,29 @@ class FileData {
   FileData({required this.path, required this.requestName});
 }
 
-class RequestApiHelperData {
+class RequestApiHelperDownloader {
+  String? name, path, url;
+
+  RequestApiHelperDownloader({this.path, this.name, this.url});
+}
+
+class RequestApiHelperDownloadData extends RequestApiHelperConfig {
+  String? path, thumnailPath, nameFile;
+  Function(dynamic)? onSuccess;
+  Function(Response)? onError;
+
+  Duration? timeout;
+  RequestApiHelperDownloadData({
+    this.path,
+    this.thumnailPath,
+    this.onSuccess,
+    this.onError,
+    this.timeout,
+    this.nameFile,
+  });
+}
+
+class RequestApiHelperData extends RequestApiHelperConfig {
   String? baseUrl;
   Map<String, dynamic>? body;
   Map<String, String>? header;